refactor: mapping 과 adapter 패키지 분리
