+++ app/src/main/java/com/techcourse/ManualHandlerMapping.java
@@ -2,7 +2,7 @@ package com.techcourse;
 
 import com.techcourse.controller.*;
 import jakarta.servlet.http.HttpServletRequest;
-import nextstep.mvc.HandlerMapping;
+import nextstep.mvc.handlermapping.HandlerMapping;
 import nextstep.mvc.controller.Controller;
 import com.techcourse.controller.ForwardController;
 import org.slf4j.Logger;
