+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -4,6 +4,8 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.handleradapter.HandlerAdapter;
+import nextstep.mvc.handlermapping.HandlerMapping;
 import nextstep.mvc.view.ModelAndView;
 
 import org.slf4j.Logger;
