+++ mvc/src/main/java/nextstep/mvc/handleradapter/AnnotationHandlerAdapter.java
@@ -1,21 +1,20 @@
-package com.techcourse;
+package nextstep.mvc.handleradapter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.HandlerAdapter;
-import nextstep.mvc.controller.Controller;
+import nextstep.mvc.handlermapping.HandlerExecution;
 import nextstep.mvc.view.ModelAndView;
+import nextstep.web.annotation.Controller;
 
-public class ManualHandlerAdapter implements HandlerAdapter {
+public class AnnotationHandlerAdapter implements HandlerAdapter {
 
     @Override
     public boolean supports(final Object handler) {
-        return handler instanceof Controller;
+        return handler.getClass().isAnnotationPresent(Controller.class);
     }
 
     @Override
     public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response, final Object handler) throws Exception {
-        final String viewName = ((Controller) handler).execute(request, response);
-        return new ModelAndView(viewName);
+        return ((HandlerExecution) handler).handle(request, response);
     }
 }
