+++ mvc/src/test/java/nextstep/mvc/controller/handlermapping/AnnotationHandlerMappingTest.java
@@ -1,9 +1,9 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.controller.handlermapping;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.AnnotationHandlerMapping;
-import nextstep.mvc.controller.HandlerExecution;
+import nextstep.mvc.handlermapping.AnnotationHandlerMapping;
+import nextstep.mvc.handlermapping.HandlerExecution;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
